feature #13_google login button fix
