+++ app/src/main/java/com/teamnexters/lastwednesday/fragment/MainFragment.java
@@ -1,13 +1,21 @@
 package com.teamnexters.lastwednesday.fragment;
 
+import android.content.Intent;
 import android.os.Bundle;
+import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.Button;
 
+import com.google.android.gms.auth.api.Auth;
+import com.google.android.gms.auth.api.signin.GoogleSignInOptions;
+import com.google.android.gms.common.ConnectionResult;
+import com.google.android.gms.common.api.GoogleApiClient;
 import com.teamnexters.lastwednesday.R;
+import com.teamnexters.lastwednesday.activity.LogInActivity;
 
 /**
  * Created by JY on 2018-01-10.
@@ -15,8 +23,10 @@ import com.teamnexters.lastwednesday.R;
  * Edited by Hyunsik on 2018-01-11.
  */
 
-public class MainFragment extends Fragment {
+public class MainFragment extends Fragment implements GoogleApiClient.OnConnectionFailedListener{
 
+    private GoogleApiClient googleApiClient;
+    private Button SignOut;
 
     public static MainFragment newInstance() {
         MainFragment fragment = new MainFragment();
@@ -36,7 +46,23 @@ public class MainFragment extends Fragment {
     @Override
     public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
         View rootView = inflater.inflate(R.layout.fragment_main, container, false);
-
+        GoogleSignInOptions signInOptions = new GoogleSignInOptions.Builder(GoogleSignInOptions.DEFAULT_SIGN_IN).requestEmail().build();
+        googleApiClient = new GoogleApiClient.Builder(getActivity()).addApi(Auth.GOOGLE_SIGN_IN_API,signInOptions).build();
+        googleApiClient.connect();
+        SignOut = (Button)rootView.findViewById(R.id.bn_logout);
+        SignOut.setOnClickListener(new View.OnClickListener(){
+            @Override
+            public void onClick(View view) {
+                Auth.GoogleSignInApi.signOut(googleApiClient);
+                Intent intent = new Intent(getActivity(), LogInActivity.class);
+                MainFragment.this.startActivity(intent);
+            }
+        });
         return rootView;
+    }
+
+    @Override
+    public void onConnectionFailed(@NonNull ConnectionResult connectionResult) {
+
     }
 }
