+++ app/src/main/res/layout/activity_main.xml
@@ -10,12 +10,4 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent" />
 
-    <Button
-        android:id="@+id/bn_logout"
-        android:layout_width="102dp"
-        android:layout_height="wrap_content"
-        android:text="Logout"
-        tools:ignore="MissingConstraints"
-        tools:layout_editor_absoluteX="266dp"
-        tools:layout_editor_absoluteY="16dp" />
 </android.support.constraint.ConstraintLayout>
