+++ app/src/main/res/layout/fragment_main.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
-<android.support.constraint.ConstraintLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
+<android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    xmlns:tools="http://schemas.android.com/tools"
     android:background="#000000">
 
     <TextView
@@ -11,4 +11,13 @@
         android:gravity="center"
         android:text="MAIN"
         android:textColor="#fff"/>
+
+    <Button
+        android:id="@+id/bn_logout"
+        android:layout_width="102dp"
+        android:layout_height="wrap_content"
+        android:text="Logout"
+        tools:ignore="MissingConstraints"
+        tools:layout_editor_absoluteX="266dp"
+        tools:layout_editor_absoluteY="16dp" />
 </android.support.constraint.ConstraintLayout>
